// redux/api/routers.ts
export enum API_ROUTES {
  LOGIN = "/patient/login",
  REGISTER = "/patient/register",
  VERIFY_OTP = "/patient/verifyOtp",
  SEND_OTP = "/patient/sendOtp",
  FIND_PATIENT_BY_ID = "/patient/fetch",
  FETCH_PATIENT_SUDIES = "/patient/study/fetch",
  UPDATE_PATIENT_PROFILE = "/patient/update",
  FILTER_PATIENTS = "/patient/patients",
  CREATE_PROTOCOL = "/protocol/create",
  FFETCH_PROTOCOLS = "/protocol/fetch",
  UPDATE_PROTOCOLS = "/protocol/update",
  DELETE_PROTOCOLS = "/protocol/delete",
  FFETCH_SURVEY = "/protocol/survey/fetch",
  CREATE_SURVEY = "/protocol/survey/create",
  UPDATE_SURVEY = "/protocol/survey/update",
  FETCH_QUESTION_TYPES = "/question/get-question-option-types",
  CREATE_SURVEY_QUESTION = "/question/create-survey-question",
  UPDATE_SURVEY_QUESTION = "/question/update-survey-question",
  FETCH_SINGLE_SURVEY_QUESTIONS = "/question/fetch",
  START_RECORD_STUDY_ACTIVITY = "/patient/study-record/save",
  FETCH_SURVEY_RECORDS_DATA = "/patient/survey-record/fetch",
  SAVE_PATIENT_RESPONSE = "/patient/patient-response/save",
  FETCH_PATIENT_RESPONSE = "/patient/patient-response/fetch",
  GET_S3_SIGNED_UPLOAD_URL = "audio/upload/getSignedUrl",
  FETCH_EMOJIS = "/emoji/fetch",
}